requirements.txt
yfinance==0.2.54
pandas>=2.0
fredapi>=0.5
python-dotenv>=1.0
requests>=2.31
